+++ src/ConfigWatcher.ts
@@ -1,7 +1,7 @@
 import * as vscode from "vscode";
 import { fixCheckSums } from "./CheckSumService";
 import { InstallStatus } from "./StickerService";
-import { attemptToInstallSticker, attemptToInstallWallpaper, getCurrentThemeAndSticker, handleInstallFailure } from "./ThemeManager";
+import { attemptToInstallSticker, attemptToInstallWallpaper, getCurrentThemeAndSticker, handleInstallFailure, handleInstallMessage, showInstallNotification } from "./ThemeManager";
 
 export const CONFIG_NAME = "doki";
 export const CONFIG_STICKER = "sticker.path";
@@ -52,16 +52,8 @@ export const watchConfigChanges = (
           handleInstallFailure(extensionContext, theme);
         } else if (hadSuccess) {
           fixCheckSums();
-          vscode.window
-            .showInformationMessage(
-              `Custom Assets installed!\n Please restart your VSCode`,
-              { title: "Restart VSCode" }
-            )
-            .then((item) => {
-              if (item) {
-                vscode.commands.executeCommand("workbench.action.reloadWindow");
-              }
-            });
+          const message = `VSCode Watermark hidden! ${handleInstallMessage}`;
+          showInstallNotification(message)
         }
         currentConfig = newBoiConfig
       })